Improve use of the test framework in the SM2 internal tests
Also general clean up of those tests

(Merged from https://github.com/openssl/openssl/pull/6386)
