+++ test/recipes/03-test_internal_sm2.t
@@ -11,9 +11,9 @@ use OpenSSL::Test;              # get 'plan'
 use OpenSSL::Test::Simple;
 use OpenSSL::Test::Utils;
 
-setup("test_sm2_sign_internal");
+setup("test_internal_sm2");
 
 plan skip_all => "This test is unsupported in a shared library build on Windows"
     if $^O eq 'MSWin32' && !disabled("shared");
 
-simple_test("test_sm2_sign_internal", "sm2_sign_internal_test", "sm2");
+simple_test("test_internal_sm2", "sm2_internal_test", "sm2");
